fix build errors

+++ mrsa_ca_rna/figures/figure_pca_predictionsXcomponents.py
@@ -38,7 +38,9 @@ def figure_setup():
     combined_df = combined.to_df()
 
     # get the MRSA outcome data to regress against
-    y_true = whole_data.obs.loc[whole_data.obs["disease"] == "MRSA", "status"]
+    y_true = whole_data.obs.loc[whole_data.obs["disease"] == "MRSA", "status"].astype(
+        int
+    )
 
     """We only have MRSA outcome data, so we have to leave out the CA data
     and truncate the combined data to MRSA data before performing regression."""
